+++ trmm_float.cu
@@ -15,21 +15,21 @@
 
 // C = alpha * A * B + beta * C
 // A is m * m col major Lower triangular, B is m * n col major, C is m * n col major
-void trmm(cublasHandle_t cublasH, int m, int n, float alpha, float *A, int lda,
-          float *B, int ldb, float beta, float *C, int ldc, int nb) {
-    int num_block = m / nb;
-    int left = m % nb;
+void trmm(cublasHandle_t cublasH, long m, long n, float alpha, float *A, long lda,
+          float *B, long ldb, float beta, float *C, long ldc, long nb) {
+    long num_block = m / nb;
+    long left = m % nb;
     float one = 1;
     cublasSgemmStridedBatched(cublasH, CUBLAS_OP_N, CUBLAS_OP_N, nb, n, nb, &alpha,
                               A, lda, nb + nb * lda, B, ldb, nb, &beta, C, ldc, nb,
                               num_block);
     if (left > 0) {
-        int offset = num_block * nb;
+        long offset = num_block * nb;
         cublasSgemm(cublasH, CUBLAS_OP_N, CUBLAS_OP_N, left, n, left, &alpha,
                     A + offset + offset * lda, lda, B + offset, ldb, &beta,
                     C + offset, ldc);
     }
-    for (int i = 1; i * nb < m; i *= 2) {
+    for (long i = 1; i * nb < m; i *= 2) {
         num_block = m / (2 * i * nb);
         left = m - (num_block * 2 * i * nb);
         cublasSgemmStridedBatched(cublasH, CUBLAS_OP_N, CUBLAS_OP_N, i * nb, n,
@@ -37,8 +37,8 @@ void trmm(cublasHandle_t cublasH, int m, int n, float alpha, float *A, int lda,
                                   2 * (i * nb + i * nb * lda), B, ldb, 2 * i * nb,
                                   &one, C + i * nb, ldc, 2 * i * nb, num_block);
         if (left > i * nb) {
-            int offset_row = i * nb + num_block * (2 * i * nb);
-            int offset_col = num_block * (2 * i * nb);
+            long offset_row = i * nb + num_block * (2 * i * nb);
+            long offset_col = num_block * (2 * i * nb);
             cublasSgemm(cublasH, CUBLAS_OP_N, CUBLAS_OP_N, left - i * nb, n, i * nb,
                         &alpha, A + offset_row + offset_col * lda, lda,
                         B + offset_col, ldb, &one, C + offset_row, ldc);
@@ -49,7 +49,7 @@ void trmm(cublasHandle_t cublasH, int m, int n, float alpha, float *A, int lda,
 int main(int argc, char *argv[]) {
     cublasHandle_t cublasH = NULL;
 
-    int m = 16384, n = 16384, nb = 512;
+    long m = 16384, n = 16384, nb = 512;
     int check = 0;
 
     if (argc >= 5) {
@@ -59,7 +59,7 @@ int main(int argc, char *argv[]) {
         check = atoi(argv[4]);
     }
 
-    int lda = m, ldb = m, ldc = m;
+    long lda = m, ldb = m, ldc = m;
 
     float *d_A = nullptr;
     float *d_B = nullptr;
